refactor(dialogs)!: move dialogs to bootstack.dialogs with _impl internals
Mirror the widgets package layout: the dialog implementations move to bootstack/dialogs/_impl/, and bootstack/dialogs/__init__.py becomes the public facade (the former widgets/dialogs.py) exposing the alert/confirm/ask_* verbs and the Dialog/FormDialog/FontDialog/ColorChooserDialog/FilterDialog classes. widgets/dialogs.py is removed. This completes the namespace curation — dialogs were the one family stuck at the awkward bootstack.widgets.dialogs path because the public name was taken by the impl package. Intra-impl and internal deep importers (dateentry, form, tableview) point at bootstack.dialogs._impl; the top-level verb import, the surface test, and the docs are repointed at bootstack.dialogs.

BREAKING CHANGE: dialog classes are imported from bootstack.dialogs (was bootstack.widgets.dialogs).
